requirements.txt
numpy
scikit-learn
pandas
gmusicapi
py-getch